Use a carousel library to make the carousel happen
