+++ package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "gsap": "^3.7.1",
     "next": "^11.1.2",
+    "pure-react-carousel": "^1.27.6",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "styled-components": "^5.3.0"
